fix(core): count only the layers that hold a KV cache
`estimate_kv_elems_per_token` multiplied its per-token figure by
`block_count` — every layer in the model. On a hybrid-attention architecture
that is wrong by the ratio of the two layer kinds:
`qwen35.full_attention_interval = 4` means only every fourth layer keeps a KV
cache, so of Qwen3.8-27B's 64 blocks just 16 do. The other 48 are linear/SSM
layers whose state is a fixed-size summary — constant in context length, not
proportional to it — so its cost belongs in a weights-side allowance.

A per-token figure is a slope, so folding those layers in over-counted them
by the entire context: 256 KiB/token instead of 64, i.e. 64 GiB rather than
16 GiB at the model's 262144-token context. Qwen3.6 declares the same key and
has carried the same defect silently.

Three surfaces read the number and were wrong together, all of them off the
one precomputed `Option<KvElemsPerToken>` on `ModelLaunchSpec`: the `kv` line
in the CLI and GUI launch banner (`launch_narration.rs:161`), the
`--cache-ram` budget and its `CacheRamHealth` band (`cache_ram.rs:116`), and
the secondary-VRAM-slot verdict (`residency.rs:96`), which can refuse a model
that in fact fits. There is exactly one production call site
(`model_catalog.rs:57`) and everything downstream receives only the
precomputed value, so the correction belongs inside the estimator and reaches
all three by construction — Tier 1, satisfied without a parity edit.

`kv_memory` already owned the hybrid concept: it reads
`full_attention_interval`, its doc already names the Qwen value, and its
fixture is already Qwen-shaped. It gains `kv_cache_layer_count`, the numeric
sibling of the `bool` it had. Division rounds up, because when the interval
does not divide the block count evenly the metadata cannot say where the
remainder falls, and one layer of over-estimate is the safe direction for a
figure the launcher plans memory against. With no such key, or an interval of
1, or a non-numeric one, `block_count` passes through untouched — every model
gglib serves today is bit-identical, and there is a named test for exactly
that rather than an assumption.

Both functions in the module now resolve their key prefix through one
`resolve_arch`, so the new one's contract is the same
`Option<&str>`-falling-back-to-`general.architecture` its sibling already had
rather than a second spelling of it. No behaviour change; its two existing
tests pin it.

Ratchet arithmetic. `kv_estimate.rs` is frozen at 341 and is still 341: the
`use` line is paid for by the module doc, which claimed the estimate "ignores
architecture-specific extras (sliding-window layers, …)" — a clause that had
to go regardless, since hybrid layers are now precisely what it does not
ignore. `kv_memory.rs` would have reached ~330 carrying the new function and
its tests, so `mod tests` moves to a `kv_memory_tests.rs` sibling via
`#[path]`, the pattern `truncation.rs` set; the module itself drops 211 to
174. No baseline was updated.

`docs/cache.md` needed no change: it describes the tiers and the
partial-memory skip rule, and carries no worked bytes-per-token example.
